IN-41 refactor: complete folders renaming
