+++ src/shared/ui/Checkbox/Checkbox.module.scss
