+++ src/shared/ui/Checkbox/Checkbox.stories.tsx
