+++ src/shared/ui/Checkbox/checkBoxSelectedIcon.tsx
