+++ src/shared/ui/Controlled/ControlledTextField.tsx
@@ -1,6 +1,6 @@
 import { Control, FieldPath, FieldValues, useController } from 'react-hook-form'
 
-import { TextField, TextFieldProps } from '../TextFieldTemp/TextField'
+import { TextField, TextFieldProps } from '../TextField/TextField'
 
 export type ControlledTextFieldProps<TFieldValues extends FieldValues> = {
   control: Control<TFieldValues>
