+++ src/shared/ui/DatePicker/DatePicker.tsx
@@ -9,10 +9,11 @@ import MultiDatePicker, { Value } from 'react-multi-date-picker'
 
 import { Calendar, CalendarOutline } from '@/src/shared/assets/icons'
 import { useTranslation } from '@/src/shared/hooks'
-import { Typography } from '@/src/shared/ui'
 import clsx from 'clsx'
 
 import s from './DatePicker.module.scss'
+
+import { Typography } from '../Typography/Typography'
 
 export type DatePickerProps = {
   className?: string
