+++ src/shared/ui/Dialog/Dialog.module.scss
