+++ src/shared/ui/ImageUploadInput/ImageUploadInput.stories.tsx
