+++ src/shared/ui/ImageUploader/ImageUploader.module.scss
