+++ src/shared/ui/Modal/Modal.module.scss
