+++ src/shared/ui/Modal/Modal.stories.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 
 import { Meta, StoryObj } from '@storybook/react'
 
-import { Button } from '../ButtonTemp/Button'
+import { Button } from '../Button/Button'
 import { Modal } from './Modal'
 
 const meta = {
