+++ src/shared/ui/ScrollArea/ScrollArea.module.scss
