+++ src/shared/ui/Spinner/Spinner.module.scss
