+++ src/shared/ui/TextArea/TextArea.stories.tsx
@@ -1,5 +1,6 @@
-import { TextArea } from '@/src/shared/ui'
 import { Meta, StoryObj } from '@storybook/react'
+
+import { TextArea } from './TextArea'
 
 const meta: Meta<typeof TextArea> = {
   argTypes: {
