+++ src/shared/ui/index.ts
@@ -1,20 +1,20 @@
 export * from './Avatar/Avatar'
-export * from './ButtonTemp/Button'
+export * from './Button/Button'
 export * from './Card/Card'
-export * from './CheckboxTemp/Checkbox'
+export * from './Checkbox/Checkbox'
 export * from './Combobox/Combobox'
 export * from './Controlled'
 export * from './DatePicker/DatePicker'
-export * from './DialogTemp/Dialog'
+export * from './Dialog/Dialog'
 export * from './Dropdown'
-export * from './ImageUploaderTemp/ImageUploader'
-export * from './ImageUploadInputTemp/ImageUploadInput'
-export * from './ModalTemp/Modal'
+export * from './ImageUploader/ImageUploader'
+export * from './ImageUploadInput/ImageUploadInput'
+export * from './Modal/Modal'
 export * from './Popover/Popover'
-export * from './ScrollAreaTemp/ScrollArea'
-export * from './SelectTemp/Select'
-export * from './SpinnerTemp/Spinner'
+export * from './ScrollArea/ScrollArea'
+export * from './Select/Select'
+export * from './Spinner/Spinner'
 export * from './Tabs'
 export * from './TextArea/TextArea'
-export * from './TextFieldTemp/TextField'
-export * from './TypographyTemp/Typography'
+export * from './TextField/TextField'
+export * from './Typography/Typography'
